move formatQS to utils
